Modify the location of three IProcesses that was generic instead of proper to a specific scale
