+++ geoindicators/src/main/groovy/org/orbisgis/BuildingIndicators.groovy
@@ -7,53 +7,6 @@ import org.orbisgis.processmanagerapi.IProcess
 
 @BaseScript Geoclimate geoclimate
 
-/**
- * This process extract geometry properties.
- *
- * @param datasource A connexion to a database (H2GIS, PostGIS, ...) where are stored the input Table and in which
- * the resulting database will be stored
- * @param inputFields Fields to conserved from the input table in the output table
- * @param inputTableName The name of the input ITable where are stored the geometries to apply the geometry operations
- * @param operations Operations that have to be applied. These operations should be in the following list (see http://www.h2gis.org/docs/dev/functions/):
- * ["st_geomtype","st_srid", "st_length","st_perimeter","st_area", "st_dimension", "st_coorddim", "st_num_geoms",
- * "st_num_pts", "st_issimple", "st_isvalid", "st_isempty"]
- * @param prefixName String use as prefix to name the output table
- *
- * @return A database table name.
- * @author Erwan Bocher
- */
-static IProcess geometryProperties() {
-return processFactory.create(
-        "Geometry properties",
-        [inputTableName: String,inputFields:String[],operations: String[]
-         , prefixName: String, datasource: JdbcDataSource],
-        [outputTableName : String],
-        { inputTableName,inputFields, operations, prefixName, datasource ->
-
-            def geometricField = "the_geom";
-            def ops = ["st_geomtype","st_srid", "st_length","st_perimeter","st_area", "st_dimension",
-                   "st_coorddim", "st_num_geoms", "st_num_pts", "st_issimple", "st_isvalid", "st_isempty"]
-
-            // The name of the outputTableName is constructed
-            String baseName = "geometry_properties"
-            String outputTableName = prefixName + "_" + baseName
-
-            String query = "CREATE TABLE $outputTableName AS SELECT "
-
-            // The operation names are transformed into lower case
-            operations.replaceAll({s -> s.toLowerCase()})
-
-            operations.each {operation ->
-                if(ops.contains(operation)){
-                    query += "$operation($geometricField) as $operation,"
-                }
-            }
-            query+= "${inputFields.join(",")} from $inputTableName"
-            logger.info("Executing $query")
-            datasource.execute query
-            [outputTableName: outputTableName]
-            }
-)}
 
 /**
  * This process extract building size properties.
