+++ geoindicators/src/main/groovy/org/orbisgis/Geoclimate.groovy
@@ -18,6 +18,8 @@ abstract class Geoclimate extends Script {
 
     public static BlockIndicators = new BlockIndicators()
 
+    public  static GenericIndicators = new GenericIndicators()
+
     public  static SpatialUnits = new SpatialUnits()
 
 }
