+++ geoindicators/src/test/groovy/org/orbisgis/BuildingIndicatorsTests.groovy
@@ -7,26 +7,6 @@ import org.junit.jupiter.api.Test
 class BuildingIndicatorsTests {
 
     @Test
-    void geometryPropertiesTest() {
-        def h2GIS = H2GIS.open([databaseName: './target/buildingdb'])
-        h2GIS.execute("""
-                DROP TABLE IF EXISTS spatial_table, test_geometry_properties;
-                CREATE TABLE spatial_table (id int, the_geom LINESTRING);
-                INSERT INTO spatial_table VALUES (1, 'LINESTRING(0 0, 0 10)'::GEOMETRY);
-        """)
-       def  p =  Geoclimate.BuildingIndicators.geometryProperties()
-       p.execute([inputTableName: "spatial_table", inputFields:["id", "the_geom"], operations:["st_issimple","st_area", "area", "st_dimension"], prefixName : "test",datasource:h2GIS])
-        assert p.results.outputTableName == "test_geometry_properties"
-        h2GIS.getTable("test_geometry_properties").eachRow {
-            row -> assert row.the_geom!=null
-                assert row.st_issimple==true
-                assertEquals(0,row.st_area)
-                assertEquals(1, row.st_dimension)
-                assertEquals(1,  row.id)
-        }
-    }
-
-    @Test
     void sizePropertiesTest() {
         def h2GIS = H2GIS.open([databaseName: './target/buildingdb'])
         String sqlString = new File(this.class.getResource("data_for_tests.sql").toURI()).text
